requirements.txt
absl-py>=0.10
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
cx-Oracle==8.3.0
cycler==0.11.0
Cython==0.29.26
dm-tree==0.1.7
flatbuffers>=1.12.0
fonttools==4.28.5
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio>=1.32.0
h5py>=2.10.0
idna==3.3
importlib-metadata==4.10.0
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.0
numpy>=1.14.5
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
#pycocotools==2.0.3
pycryptodome==3.12.0
pydotplus==2.0.2
PyOpenGL==3.1.5
pyparsing==3.0.6
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
python-dateutil==2.8.2
pytz==2021.3
qimage2ndarray==1.8.3
requests>=1.9.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn>=1.1.1
scipy==1.8.1
seaborn==0.11.2
six>=1.6.1
tensorflow==2.4.2
tensorflow-estimator==2.4.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions>=3.7.4
urllib3==1.26.7
Werkzeug==2.0.2
wrapt>=1.12.1
zipp==3.6.0
